# Conflicts:
#	clients/src/main/resources/i18n/message.properties
#	customer-service/src/main/java/com/unicoin/customer/entity/User.java
#	customer-service/src/main/java/com/unicoin/customer/ex/ExceptionCode.java
#	customer-service/src/main/java/com/unicoin/customer/service/impl/UserServiceImpl.java

+++ customer-service/src/main/java/com/unicoin/customer/service/UserService.java
@@ -16,7 +16,7 @@ public interface UserService {
 
     JwtResponse login();
 
-    void updateCustomer(Integer id  , AddCustomerForm addCustomerForm);
+    void updateCustomer(Long id  , AddCustomerForm addCustomerForm);
 
     void uDeleteCustomer(String phoneNumer);
